+++ Packages/com.nidonocu.vrcunitytoolbar/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+### 2.1.0
+* Confirmed compatible with VRChat SDK 3.2.0
+* Fixed changing of anchored transform values when smart duplicating Unity UI objects
+
 ### 2.0.1
 * Fixes a bug when compiling projects